% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate.nelson.1965.R
\name{cate.nelson.1965}
\alias{cate.nelson.1965}
\title{Cate & Nelson Quadrants Analysis, 1965}
\usage{
cate.nelson.1965(data, STV, RY, target)
}
\arguments{
\item{data}{argument to call a data.frame or data.table containing the data}

\item{STV}{argument to call the vector or column containing the soil test value (STV) data}

\item{RY}{argument to call the vector or column containing the relative yield (RY) data}

\item{target}{argument to specify the RY target (numeric) to estimate the critical STV}
}
\value{
it returns an object of type list containing the main results plus a ggplot object with the figure display
}
\description{
This function runs the quadrants analysis suggested by Cate and Nelson (1965)
}
\details{
See Cate, R. B. Jr., and L. A. Nelson. 1965. A rapid method for correlation of soil test analysis
with plant response data. North Carolina Agric. Exp. Stn., International soil Testing Series l. No. 1.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{eval_tidy}},\code{\link[rlang]{defusing-advanced}}
\code{\link[stats]{lm}},\code{\link[stats]{anova}}
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{scale_manual}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{annotate}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{theme}}
}
