% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate_nelson_1971.R
\name{cate_nelson_1971}
\alias{cate_nelson_1971}
\title{Cate & Nelson Quadrants Analysis, 1971}
\usage{
cate_nelson_1971(data = NULL, stv, ry, tidy = FALSE, plot = FALSE)
}
\arguments{
\item{data}{argument to call a data.frame or data.table containing the data}

\item{stv}{argument to call the vector or column containing the soil test value (stv) data}

\item{ry}{argument to call the vector or column containing the relative yield (ry) data}

\item{tidy}{logical operator (TRUE/FALSE) to decide the type of return. TRUE returns a data.frame, FALSE returns a list (default).}

\item{plot}{logical operator (TRUE/FALSE) to decide the type of return. TRUE returns a ggplot,
FALSE returns either a list (tidy == FALSE) or a data.frame (tidy == TRUE).}
}
\value{
it returns an object of type \code{ggplot} when plot == TRUE. If plot == FALSE, it returns an object of type \code{data.frame} if tidy == TRUE, otherwise, it returns a list

it returns an object of type list containing the main results plus a ggplot object with the figure display
}
\description{
This function runs the quadrants analysis suggested by Cate and Nelson (1965)
}
\details{
See Cate, R. B. Jr., and L. A. Nelson. 1965. A rapid method for correlation of soil test analysis
with plant response data. North Carolina Agric. Exp. Stn., International soil Testing Series l. No. 1.
This code was adapted from Mangiafico, S. S. (2013). Cate-Nelson Analysis for Bivariate Data Using R-project.
The Journal of Extension, 51(5), Article 33. https://tigerprints.clemson.edu/joe/vol51/iss5/33
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{eval_tidy}},\code{\link[rlang]{defusing-advanced}}
\code{\link[stats]{lm}},\code{\link[stats]{anova}}
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{scale_manual}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{annotate}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{theme}}
}
