% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freitas.1966}
\alias{freitas.1966}
\title{Dataset 2.}
\format{
this data frame has 24 rows and the following 2 columns:
\describe{
\item{RY}{relative yield, \%}
\item{STK}{soil test potassium, ppm}
}
}
\source{
\url{https://doi.org/10.2136/sssaj1971.03615995003500040048x}
}
\usage{
freitas.1966
}
\description{
Example dataset containing real data reported by Cate & Nelson (1971) from Freitas et al. (1966).
Soil test potassium values (STK) and relative yield as percentage (RY).
}
\keyword{datasets}
