% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freitas1966}
\alias{freitas1966}
\title{Dataset 2}
\format{
this data frame has 24 rows and the following 2 columns:
\describe{
\item{RY}{relative yield, \%}
\item{STK}{soil test potassium, ppm}
}
}
\source{
Freitas et al. (1966) cited and used by Cate & Nelson (1971). Soil Sci. Soc. Am. Proc. 35:658-659
}
\usage{
freitas1966
}
\description{
Example dataset containing real data reported by Cate & Nelson (1971) from Freitas et al. (1966).
Soil test potassium values (STK) and relative yield as percentage (RY).
}
\keyword{datasets}
