% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_plateau.R
\name{linear_plateau}
\alias{linear_plateau}
\alias{LP_f}
\alias{SS_LP}
\title{Linear Plateau}
\usage{
LP_f(x, intercept, slope, Xc)

SS_LP(x, intercept, slope, Xc)

linear_plateau(data = NULL, STV, RY, resid = FALSE, plot = FALSE)
}
\arguments{
\item{data}{Optional argument to call and object of type data.frame or data.table
containing the STV and RY data, Default: NULL}

\item{STV}{name of the vector containing soil test values (-) of type \code{numeric}.}

\item{RY}{name of the vector containing relative yield values (\%) of type \code{numeric}.}

\item{resid}{boolean (TRUE/FALSE) to plot residuals analysis, Default: FALSE}

\item{plot}{boolean (TRUE/FALSE) to plot the linear-plateau model, Default: FALSE
Default: FALSE}
}
\value{
returns a \code{data.frame} if plot = FALSE, if plot = TRUE

LP_f: vector of the same length as x using the linear-plateau function

SS_LP: selfStart object to pass into the linear_plateau fit

linear_plateau: function
}
\description{
This function helps to fit a linear-plateau model in order to
estimate critical soil test values (CSTV)
}
\details{
This functions fits a linear-plateau model using selfStart initial values
}
\examples{
\dontrun{
if(interactive()){
 # Example dataset
 dat <- data.frame("ry" = c(65,80,85,88,90,94,93,96,97,95,98,100,99,99,100),
                   "stv" = c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15))
 # Run
 fit_example_lp <- linear_plateau(data = dat, RY = ry, STV = stv, resid = TRUE, plot = FALSE)
 fit_example_lp
 }
}
}
\seealso{
\code{\link[rlang]{eval_tidy}},\code{\link[rlang]{defusing-advanced}}
\code{\link[minpack.lm]{nlsLM}}
\code{\link[stats]{AIC}},\code{\link[stats]{lm}},\code{\link[stats]{optim}},\code{\link[stats]{coef}},\code{\link[stats]{predict}}
\code{\link[AICcmodavg]{AICc}}
\code{\link[modelr]{model-quality}}
\code{\link[nlstools]{nlsResiduals}}
\code{\link[tidyr]{reexports}}
\code{\link[dplyr]{bind}}
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_rug}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{geom_path}},\code{\link[ggplot2]{annotate}},\code{\link[ggplot2]{scale_continuous}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{theme}}
\code{\link[ggpp]{annotate}}
}
