% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_alcc.R
\name{mod_alcc}
\alias{mod_alcc}
\title{Modified Arcsine-Log Calibration Curve}
\usage{
mod_alcc(data = NULL, RY, STV, target, confidence, tidy = FALSE, plot = FALSE)
}
\arguments{
\item{data}{Optional argument to call and object of type data.frame or data.table
containing the STV and RY data, Default: NULL}

\item{RY}{name of the vector containing relative yield values (\%) of type \code{numeric}.}

\item{STV}{name of the vector containing soil test values (-) of type \code{numeric}.}

\item{target}{\code{numeric} value of relative yield target (e.g. 90 for 90\%) to estimate the CSTV.}

\item{confidence}{\code{numeric} value of confidence level (e.g. 0.95 for
significance = 0.05)}

\item{tidy}{logical operator (TRUE/FALSE) to decide the type of return. TRUE returns a data.frame, FALSE returns a list (default).}

\item{plot}{logical operator (TRUE/FALSE) to decide the type of return. TRUE returns a ggplot,
FALSE returns either a list (tidy == FALSE) or a data.frame (tidy == TRUE).}
}
\value{
an object of type \code{data.frame} if tidy == TRUE, otherwise, it returns a list
}
\description{
This function runs the modified arcsine-log calibration curve to
estimate critical soil test values (CSTV) following Correndo et al. (2017)
}
\details{
See Correndo et al. (2017)
}
\examples{
\dontrun{
if(interactive()){
 # Example 1 dataset
 dat <- data.frame("ry" = c(65,80,85,88,90,94,93,96,97,95,98,100,99,99,100),
                   "stv" = c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15))
 # Run
 fit_example <- mod_alcc(data = dat, RY = ry, STV = stv, target=90, confidence = 0.95)
 fit_example
 }
}
}
\seealso{
\code{\link[rlang]{eval_tidy}},\code{\link[rlang]{defusing-advanced}}
\code{\link[stats]{TDist}},\code{\link[stats]{cor}},\code{\link[stats]{cor.test}},\code{\link[stats]{sd}}
\code{\link[dplyr]{bind}},\code{\link[dplyr]{filter}}
\code{\link[tidyr]{nest}}
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{scale_manual}},\code{\link[ggplot2]{geom_rug}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{geom_path}},\code{\link[ggplot2]{scale_continuous}},\code{\link[ggplot2]{annotate}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{theme}}
\code{\link[ggpp]{annotate}}
}
