% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltestcorr-package.R
\docType{package}
\name{soiltestcorr-package}
\alias{soiltestcorr}
\alias{soiltestcorr-package}
\title{soiltestcorr: Soil Test Correlation and Calibration in R}
\description{
The functions offered by this package are designed to assist users on the correlation analysis of crop yield and soil test values. All functions estimate critical soil test values using alternative approaches such as: 1) the modified arcsine-log calibration curve (alcc); 2) the graphical Cate-Nelson quadrants-analysis, 3) the statistical Cate-Nelson quadrants-analysis, 4) the linear-plateau regression, 5) the quadratic-plateau regression, and 6) the Mitscherlich regression.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adriancorrendo.github.io/soiltestcorr/}
  \item Report bugs at \url{https://github.com/adriancorrendo/soiltestcorr/issues}
}

}
\author{
\strong{Maintainer}: Adrian A. Correndo \email{correndo@ksu.edu} (\href{https://orcid.org/0000-0002-4172-289X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Adrian A. Correndo \email{correndo@ksu.edu} (\href{https://orcid.org/0000-0002-4172-289X}{ORCID})
  \item Austin Pearce \email{awpearc2@ncsu.edu} (\href{https://orcid.org/0000-0002-2541-896X}{ORCID})
  \item Ignacio A. Ciampitti \email{ciampitti@ksu.edu} (\href{https://orcid.org/0000-0001-9619-5129}{ORCID})
}

}
\keyword{internal}
