% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_test}
\alias{data_test}
\title{Dataset 1}
\format{
this data frame has 137 rows and the following 2 columns:
\describe{
\item{STV}{soil test value}
\item{RY}{relative yield, \%}
}
}
\source{
\doi{10.7910/DVN/NABA57}
}
\usage{
data_test
}
\description{
Example dataset containing hypothetical pairs of soil test value (STV)
and relative yield (RY).
}
\keyword{datasets}
