% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_plateau.R
\name{linear_plateau}
\alias{linear_plateau}
\alias{LP_f}
\alias{SS_LP}
\title{Linear-plateau response function}
\usage{
LP_f(x, intercept, slope, cx)

SS_LP(x, intercept, slope, cx)

linear_plateau(
  data = NULL,
  stv,
  ry,
  target = NULL,
  tidy = FALSE,
  plot = FALSE,
  resid = FALSE
)
}
\arguments{
\item{x}{selfstart vector for independent variable, Default: NULL}

\item{intercept}{selfstart arg. for intercept Default: NULL}

\item{slope}{selfstart arg. for slope Default: NULL}

\item{cx}{selfstart arg. for critical X (cx) value Default: NULL}

\item{data}{Optional argument to call and object of type data.frame or data.table
containing the soil test value (STV) and relative yield (RY) data, Default: NULL}

\item{stv}{name of the vector containing soil test values (-) of type \code{numeric}.}

\item{ry}{name of the vector containing relative yield values (\%) of type \code{numeric}.}

\item{target}{\code{numeric} value of relative yield target (e.g. 90 for 90\%) to estimate the CSTV.
The target needs to be < plateau, otherwise, target = plateau.}

\item{tidy}{logical operator (TRUE/FALSE) to decide the type of return. TRUE returns a data.frame, FALSE returns a list (default).}

\item{plot}{logical operator (TRUE/FALSE) to plot the linear-plateau model, Default: FALSE}

\item{resid}{logical operator (TRUE/FALSE) to plot residuals analysis, Default: FALSE}
}
\value{
returns an object of type \code{ggplot} if plot = TRUE.

returns a residuals plot if resid = TRUE.

returns an object of class \code{data.frame} if tidy = TRUE,

returns an object of class \code{list} if tidy = FALSE.

LP_f: vector of the same length as x using the linear-plateau function

SS_LP: selfStart object to pass into the linear_plateau fit

linear_plateau: function
}
\description{
This function helps to fit a linear-plateau model in order to
estimate critical soil test values (CSTV) above which yield response becomes flat.
}
\details{
See \href{https://adriancorrendo.github.io/soiltestcorr/articles/linear_plateau_tutorial.html}{online-documentation} for additional details.
}
\note{
For extended reference, we recommend to visit:
\url{https://gradcylinder.org/linear-plateau/} & \url{https://github.com/austinwpearce/SoilTestCocaCola} by Austin Pearce.
Self-start function code adapted from nlraa package by F. Miguez \url{https://github.com/femiguez/nlraa}
}
\examples{
\donttest{
 # Example dataset
 dat <- data.frame("ry" = c(65,80,85,88,90,94,93,96,97,95,98,100,99,99,100),
                   "stv" = c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15))
 # Run
 fit_example_lp <- linear_plateau(data = dat, 
 ry = ry, stv = stv, resid = TRUE, plot = FALSE)
 fit_example_lp
}
}
\references{
Anderson, R. L., and Nelson, L. A. (1975).
A Family of Models Involving Intersecting Straight Lines and Concomitant Experimental Designs Useful in Evaluating Response to Fertilizer Nutrients.
\emph{Biometrics, 31(2), 303–318.} \doi{10.2307/2529422}
}
\seealso{
\code{\link[rlang]{eval_tidy}},\code{\link[rlang]{defusing-advanced}}
\code{\link[minpack.lm]{nlsLM}}
\code{\link[stats]{AIC}},\code{\link[stats]{lm}},\code{\link[stats]{optim}},\code{\link[stats]{coef}},\code{\link[stats]{predict}}
\code{\link[AICcmodavg]{AICc}}
\code{\link[modelr]{model-quality}}
\code{\link[nlstools]{nlsResiduals}}
\code{\link[tidyr]{reexports}}
\code{\link[dplyr]{bind}}
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_rug}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{geom_path}},\code{\link[ggplot2]{annotate}},\code{\link[ggplot2]{scale_continuous}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{theme}}
\code{\link[ggpp]{annotate}}
}
