% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltestcorr-package.R
\docType{package}
\name{soiltestcorr-package}
\alias{soiltestcorr}
\alias{soiltestcorr-package}
\title{soiltestcorr: Soil Test Correlation and Calibration in R}
\description{
The soiltestcorr-package is designed to assist users on the correlation analysis of crop yield and soil test values. The soiltestcorr package estimates critical soil test values using R several different approaches such as: 1) the modified Arcsine-log calibration curve model originally described by Dyson and Conyers (2013) and modified by Correndo et al. (2017); 2) the graphical Cate & Nelson approach (Cate & Nelson, 1965), 3) the statistical Cate & Nelson approach (Cate & Nelson, 1965), 4) linear-plateau regression, 5) quadratic-plateau regression, and 6) Mitscherlich regression.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adriancorrendo.github.io/soiltestcorr/}
}

}
\author{
\strong{Maintainer}: Adrian A. Correndo \email{correndo@ksu.edu} (\href{https://orcid.org/0000-0002-4172-289X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Adrian A. Correndo \email{correndo@ksu.edu} (\href{https://orcid.org/0000-0002-4172-289X}{ORCID})
  \item Austin Pearce \email{awpearc2@ncsu.edu} (\href{https://orcid.org/0000-0002-2541-896X}{ORCID})
  \item Ignacio A. Ciampitti \email{ciampitti@ksu.edu} (\href{https://orcid.org/0000-0001-9619-5129}{ORCID})
}

}
\keyword{internal}
